feat: add support for depth parameter in orderbook query
This commit adds new method to DLOBClient, which allows to request
orderbook with a required depth. To handle cases when depth is not
provided, default constant can be used, which is set to value of 10.

Motivation: currently DLOBClient requests the entirety of orderbook,
which contains around 50 levels. Sometimes this can be unecessary, for
example if we are only interested in best bid/ask, which is why it can
be useful to have an option to control the depth of requested orderbook.